requirements.txt
flask
line-bot-sdk
yfinance
python-dotenv
gunicorn 